// src/app/shared/types/api-types/bire-version-measure-dto.interface.ts
export interface BireVersionMeasureDTO {
  counterCode?: string;
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  birFamilyCode?: string;
  birStructureType?: string;
  ratingCode?: string;
  referenceValue?: string;
  limitDate?: Date;
  versionNumber?: string;
}